Make payment number series configurable
Move payment sending to a service class because there are already two
dependencies to inject.
