+++ service/src/main/kotlin/fi/espoo/evaka/payments/Payments.kt
@@ -11,7 +11,6 @@ import fi.espoo.evaka.reports.getServiceVoucherReport
 import fi.espoo.evaka.shared.DaycareId
 import fi.espoo.evaka.shared.EvakaUserId
 import fi.espoo.evaka.shared.PaymentId
-import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.db.DatabaseEnum
 import fi.espoo.evaka.shared.domain.BadRequest
@@ -102,54 +101,3 @@ data class Payment(
     val sentAt: HelsinkiDateTime?,
     val sentBy: EvakaUserId?,
 )
-
-fun sendPayments(
-    tx: Database.Transaction,
-    now: HelsinkiDateTime,
-    integrationClient: PaymentIntegrationClient,
-    user: AuthenticatedUser.Employee,
-    paymentIds: List<PaymentId>,
-    paymentDate: LocalDate,
-    dueDate: LocalDate
-) {
-    val payments = tx.readPaymentsByIds(paymentIds)
-
-    val notDrafts = payments.filterNot { it.status == PaymentStatus.DRAFT }
-    if (notDrafts.isNotEmpty()) {
-        throw BadRequest("Some payments are not drafts")
-    }
-
-    val units = tx.getPaymentUnitsByIds(payments.map { it.unit.id }).associateBy { it.id }
-
-    var nextPaymentNumber = tx.getMaxPaymentNumber().let { if (it >= 5000000000) it + 1 else 5000000000 }
-    val updatedPayments = payments.flatMap { payment ->
-        val unit = units[payment.unit.id] ?: throw Error("Unit ${payment.unit.id} not found")
-
-        // Skip payments whose unit has missing payment details
-        val missingDetails = listOf(unit.name, unit.businessId, unit.iban, unit.providerId).any { it.isNullOrBlank() }
-        if (missingDetails) {
-            logger.warn { "Skipping payment ${payment.id} because unit ${payment.unit.id} has missing payment details" }
-            return@flatMap listOf()
-        }
-
-        val updatedPayment = payment.copy(
-            number = nextPaymentNumber,
-            paymentDate = paymentDate,
-            dueDate = dueDate,
-            unit = payment.unit.copy(
-                name = unit.name,
-                businessId = unit.businessId,
-                iban = unit.iban,
-                providerId = unit.providerId,
-            ),
-            sentBy = user.evakaUserId
-        )
-        nextPaymentNumber += 1
-        listOf(updatedPayment)
-    }
-
-    val sendResult = integrationClient.send(updatedPayments)
-    logger.info { "Successfully sent ${sendResult.succeeded.size} payments: ${sendResult.succeeded.map { it.id }.joinToString(", ")}" }
-    logger.info { "Failed to send ${sendResult.failed.size} payments: ${sendResult.failed.map { it.id }.joinToString(", ")}" }
-    tx.updatePaymentDraftsAsSent(sendResult.succeeded, now)
-}
